Create Longest Harmonious Subsequence
Problem statement
You are given an array ‘ARR’ of 'N' integers. Your task is to find the longest harmonious subsequence of this array.

A sequence is Harmonious if the difference between the maximum and the minimum of all the elements is exactly 1.

For example
[3, 4, 3, 3, 3, 3, 4] is a harmonic array as the maximum of all the elements is 4 and minimum of all the elements is 3. So, the difference between the maximum and the minimum = 4 - 3 = 1.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1<= T <= 100
1 <= N <= 10^4
1 <= ARR[i] <= 10^8

Time Limit: 1 sec
Sample Input 1:
2
4
1 2 2 1
4
1 2 3 4
Sample Output 1:
4
2
Explanation for Sample Input 1:
In the first test case, the given array is [1, 2, 2, 1]. If we take the complete array, then the maximum of all the elements is 2 and the minimum of all the elements is 1. So, the difference between the maximum and the minimum = 2 - 1 = 1. Hence, the longest Harmonic subsequence is [1, 2, 2, 1] and its length is 4.

In the second test case, the given array is [1, 2, 3, 4]. If we take the complete array, then the maximum of all the elements is 4 and the minimum of all the elements is 1. 
So, the difference between the maximum and the minimum = 4 - 1 = 3. So, it is not a Harmonic subsequence. If we take subsequence as [1, 2], then the maximum of all the elements is 2 and the minimum of all the elements is 1. 
So, the difference between the maximum and the minimum = 2 - 1 = 1. So, it is a harmonic subsequence and this is the longest harmonic subsequence. If we take subsequence as [2, 3] or [3, 4], then we will get the same answer.
Sample Input 2:
2
2
1 2
10
1 2 2 3 4 5 1 1 1 1
Sample Output 2:
2
7