Create K'th Special Number in Range
Problem statement
Ninja wants to practice martial arts in order to master it but he has a pending school assignment and wants your help to solve it.

You are given two integers ‘L’ and ‘R’ denoting the range of integers [L, R]. A special number is a number that has (101)2 subarray in its binary representation. You have a simple task, you need to find the ‘K’th’ special number that lies in the given range, print -1 if it is not possible.

For Example :
If L = 5 and R = 7, and you need to find the 1’st special number lying in the range [5, 7]

The answer for this is equal to 5, as 5 is the first special number lying in the given range, it is a special number because it has a subarray 101 in its binary representation.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 ≤ T ≤ 10      
1 ≤ L ≤ R ≤ 50000
1 ≤ K ≤ 50000

Time limit: 1 sec
Sample Input 1 :
2
5 7 1
8 11 3
Sample Output 1 :
5
-1
Explanation For Sample Input 1 :
For test case 1 :
We will print 5 because:
5 is the first special number lying in the given range, it is a special number because it has a subarray 101 in its binary representation.

For test case 2 : 
We will print -1 because:
There are two special numbers lying in the range [8, 11], these are 10 = (1010)2 and 11 = (1011)2 as both of them have a subarray of 101 in their binary representation. As third special number doesn’t exist in this range, hence we will print -1.
Sample Input 2 :
2
1 10 2
1 5 1
Sample Output 2 :
10
5