Create Stones falling
Problem statement
This time Ninja has a 'GRID' of size 'N' x 'M' of binary values (0s and 1s) the 1s in a cell represent the stones. A stone will not fall when that satisfies 2 conditions:

1) It is directly connected to the top of the grid

2) At least one of its adjacent (top,bottom,left, right) stones will not fall.

Ninja has 'R' sequence of locations of stones (given in a matrix 'REM' of size 'R' X 2). In each case, he wants to remove the stone placed at (i, j) place but due to this action, some other stones will also fall. So he needs your help to find the number of stones that fall after each removal in sequence.

Return an array of size 'R' denoting the number of stones that fall after each removal of a stone.

Note:

Sometimes a removal may refer to a location with no stone, in that case, no stone fall.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 5
1 <= N, M <= 200
1 <= R <= N * M   

Where ‘T’ is the number of test cases, ‘N’ and ‘M’ is the number of rows and columns respectively and ‘R’ is the number of removal operations.

Time limit: 1 sec.
Sample Input 1:
2
3 3 1
1 1 0 
1 0 0
1 1 1
1 0
2 3 2
1 1 1
1 0 1
0 0
1 2
Sample Output 1:
3
1 0    
Explanation of Sample Output 1:
Test Case 1 : On removing stone at (1, 0) all the stones below it i.e. (2, 0), (2, 1), (2, 2) will fall.
Hence, the answer is 3 

Test Case 2:  Here we have two locations to remove.
First is (0, 0), on removing this, stone present at (1, 0) will fall so the answer for this location is 1.
In the second location, there is no stone available so we return 0. 
Sample Input 2:
1
3 3 2
1 0 1
0 1 1
0 1 0
1 2
1 1
Sample Output 2:
2 0