Create Minimum Deletions To Make Character Frequencies Unique
Problem statement
You are given a string ‘STR’ consisting of lowercase English alphabets. Your take is to find out the minimum number of character deletions required such that each character in the final string has a unique frequency.

Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 10
2 <= |STR| <= 10^5

Where |STR| denotes the length of the initial string ‘STR’. 

Time limit: 1 second
Sample Input 1 :
2
babccbc
gttqtq
Sample Output 1 :
1
0
Explanation For Sample Input 1 :
Test Case 1 :
The frequency of ‘a’ is 1, of ‘b’ is 3 and of ‘c’ is 3. Therefore after deletion of one character of either ‘b’ or ‘c’ makes the frequencies 1,2 and 3.

Test Case 2 :
The frequency of 'g' is 1, 't' is 3 and 'q' is 2. Therefore there is no need for a delete operation.
Sample Input 2 :
2
abbccc
cbbd
Sample Output 2 :
0
1
Explanation For Sample Input 2 :
For the first test case, each character already has a unique frequency(1 for ‘a’, 2 for ‘b’ and 3 for ‘c’), hence we don’t need to do any deletions.

For the second test case frequency of ‘c’ and ‘d’ is equal hence we need to delete either of them. Therefore the answer is 1.