Create Find MSB In O(1)
Problem statement
You are given a positive integer 'N'. Your task is to find the greatest integer less than or equal to 'N' which is a power of 2.

For Example:
If N = 14, then the nearest integer that is less than or equal to 14 and is a power of two is 8(2^3). So, the answer is 8.
Follow Up:
Can you solve this in constant time and space complexity?
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 2 * 10^3
1 <= N <= 10^9

Time Limit: 1 second
Sample Input 1:
2
4
22
Sample Output 1:
4
16
Explanation for sample 1:
For the first test case, 4 itself is a power of two.
For the second test case, the nearest integer that is less than or equal to 22 and also is a power of two is 16.
Sample Input 2:
2
1
63
Sample Output 2:
1
32
Explanation for sample 1:
For the first test case, 1 itself is a power of two.
For the second test case, the nearest integer that is less than or equal to 63 and also is a power of two is 32.